requirements.txt
aiohttp==3.7.2
aiohttp-session==2.9.0
aioredis==1.3.1
alembic==1.4.3
asgiref==3.3.1
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.3
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
distro-info===0.18ubuntu0.18.04.1
Django==3.1.3
django-sslserver==0.22
facebook-sdk==3.1.0
Flask==1.0.2
Flask-BabelEx==0.9.4
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.1.0
h11==0.12.0
hiredis==1.1.0
hyperlink==21.0.0
idna==2.6
idna-ssl==1.1.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
korbit-python==0.6.0
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==1.0.2
multidict==5.0.0
paramiko==2.7.2
passlib==1.7.1
pgadmin4==4.18
psutil==5.5.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.10.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
pyxdg==0.25
requests==2.25.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
soupsieve==2.1
speaklater==1.3
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
sshtunnel==0.2.2
tornado==6.1
twilio==6.52.0
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
unattended-upgrades==0.1
urllib3==1.26.2
uvicorn==0.13.3
uWSGI==2.0.19.1
validate-email==1.3
websocket==0.2.1
websocket-client==0.56.0
websockets==8.1
Werkzeug==1.0.1
WTForms==2.2.1
yarl==1.6.2
zope.event==4.5.0
zope.interface==5.2.0
